refactor: Handle test API clients in `Infrastructure`
