+++ api/tests/common/infra.rs
@@ -1,10 +1,14 @@
+use rocket::local::asynchronous::Client;
 use rstest::fixture;
+use wiremock::MockServer;
 
 use cron_mon_api::infrastructure::database::{run_migrations, DbPool};
 use cron_mon_api::infrastructure::models::{
     api_key::ApiKeyData, job::JobData, monitor::MonitorData,
 };
+use cron_mon_api::rocket;
 
+use super::auth::setup_mock_jwks_server;
 use super::postgres::seed_db;
 use super::seeds::{api_key_seeds, job_seeds, monitor_seeds};
 use super::{postgres_container, PostgresContainer};
@@ -17,13 +21,17 @@ pub async fn infrastructure() -> Infrastructure {
 pub struct Infrastructure {
     pub container: PostgresContainer,
     pub pool: DbPool,
+    mock_server: Option<MockServer>,
 }
 
+/// A test helper struct to create infrastructure to support integration tests.
 impl Infrastructure {
+    /// Create a new, default instance of Infrastructure.
     pub async fn create() -> Self {
         Self::new(monitor_seeds(), job_seeds(), api_key_seeds()).await
     }
 
+    /// Create a new instance of Infrastructure with the provided seeds.
     pub async fn from_seeds(
         monitor_seeds: Vec<MonitorData>,
         job_seeds: Vec<JobData>,
@@ -44,6 +52,21 @@ impl Infrastructure {
         // See data seeds for the expected data (/api/tests/common/mod.rs)
         let pool = seed_db(&monitor_seeds, &job_seeds, &api_key_seeds).await;
 
-        Self { container, pool }
+        Self {
+            container,
+            pool,
+            mock_server: None,
+        }
+    }
+
+    /// Retrieve a test client for the API, linked to this Infrastructure.
+    pub async fn test_api_client(&mut self, kid: &str) -> Client {
+        self.mock_server = Some(setup_mock_jwks_server(kid).await);
+
+        let client = Client::tracked(rocket())
+            .await
+            .expect("Invalid rocket instance");
+
+        client
     }
 }
